[Tracking] Add client side plugin.
Add enginePlugin.js with Tracker MVC.

Modify ShapeModel: Each position has the property byMachine.
If this is true,  this means the position created by the tracker and
not the user.  This is used to let the tracker later change positions
of shapes again until the next manually inserted key frame.