Revert Open3D to 0.11.2
This change fixes a dependency conflict that causes a very old version
of Jinja2 to be installed (see #4618). Open3D is only used for PCD
conversion, so this reverting to an older version is acceptable.

This is a temporary workaround. As a future update we'd like to get rid
of the Open3D dependency, since it pulls in a bunch of unnecessary
dependencies. Also, we'd like to make python dependency resolution more
robust by pinning all (recursive) dependencies to specific versions with
a tool like pip-tools or poetry.