Add a label constraint to the treafik container in docker-compose
This ensures that, when multiple instances of CVAT are running through
docker-compose, their treafik instances don't both try to apply routing
for eachother. This is done by applying a label to each CVAT container,
and constraining the treafik instance based on that label.

The label was chosen to be the combination of CVAT_HOST and
CVAT_EXT_PORT. CVAT_EXT_PORT is a new environment variable which sets
the port the service is exposed as.

I also commented out ports that aren't strictly necessary; this,
however, is a breaking change, so may be reverted.