Remove `--` from some `yarn run` commands
In Yarn 1.0, the `--` is no longer required, and in Yarn 2.0 it is now
passed to the script verbatim, thus breaking the command.

There are also usages of `--` in `lint-staged.config.js`, but in those
cases, passing `--` to the script would be mildly beneficial (ensuring that
none of the file names are interpreted as options), so I'm keeping those.