cvat-sdk: remove broken and unused body-to-params conversion
It's broken, because it generates a multipart/form-data body that Django
REST Framework cannot parse.  DRF doesn't examine the individual
Content-Type headers of parts, so it interprets the JSON-encoded values as
`str()`-encoded values, which corrupts them.

Since this logic is also currently unused (and it's unclear whether it will
ever be used again), remove it.