Replace unintended Cyrillic C letters with Latin C
The only change in production code is the one in `shortcuts-reducer.ts`.  It
makes no functional difference, however, because the shortcut in question is
not handled by the regular shortcut mechanism; the control key is handled
separately in `interactionHandler.ts`, while the string "ctrl" in
`shortcuts-reducer.ts` is only used for display purposes.

Fix ESLint errors in altered files, as well.