SDK: add a utility function for atomically writing a file
This functionality is already used in the SDK, and I plan to use it in the
PyTorch dataset adapter, so it's useful to factor it out into a dedicated
function.

The new function also improves over the original implementation in that it
will try multiple temporary file names, and thus it will not fail if the
temporary file is already present (e.g.  if it was left there by a crashed
process).