Partially integrate Django REST Framework into cvat/apps/dataset_repo
The main goal is to align authentication logic between `dataset_repo` and
the rest of the API. In particular, the common logic disables CSRF checks
for queries that have token authentication, which will allow the Git API
to be used from the SDK.

The interactive HTML API pages that DRF provides are pretty nice too.

For one of the endpoints, fix the UI to pass the correct Content-Type,
otherwise, DRF rejects the request.