Convert the "services" fixture into sessionstart/sessionfinish hooks
This makes the test output less messy. Fixtures are only created after the
first test starts running, so `docker-compose` output becomes intermixed
with the test log. The hooks, on the other hand, run before the tests are
even collected, so the output is separated from the test log.

This also makes runs with options like `--stop-services` a bit faster,
because pytest doesn't have to waste time collecting tests that won't even
be run.