TaskVisionDataset: add a way to specify a custom label-name-to-index mapping
The default numbering is generally unpredictable, so it's not especially
useful outside of basic use cases. Given that there's no way to specify
custom indexes for labels on the server side, make it possible on the client
side.

The input dictionary maps label _names_ to indexes, because names is what a
CVAT user would be familiar with. The output dictionary in the `Target`
class still maps label _IDs_, because that's what CVAT uses in the
annotation data models.