tests: fix long shutdown times for the webhook_receiver container
Currently, `server.py` runs as PID 1, which means that it won't be
terminated by a `SIGTERM` signal unless it explicitly handles it (which it
doesn't). So when Docker tries to shut the container down, it sends the
server a `SIGTERM`, which gets ignored, and then sits there for 10 seconds
before sending it a `SIGKILL`.

To work around this, enable the built-in Docker init program, which forwards
signals to the Python server. Since the Python server is no longer PID 1,
`SIGTERM` will now shut it down immediately.

The `init` option is supported starting from the Compose format version 3.7,
so bump the version.