SDK: fix field serialization for multipart/form-data requests
Django REST Framework ignores the Content-Type on request body parts, so
it doesn't know that they are JSON-encoded. Instead, it just tries to decode
each part as if it was an `str()`-encoded value.

Change the encoding to match the decoding. The only type this matters for
is `str`, because `json.dumps` and `str` produce different encodings for
`str` values.

Remove `none_type` from the list of encodable types since, to my knowledge,
there's no way to encode a `None` value as a `multipart/form-data` part in
a way that DRF will understand.