Fix a CI issue that causes unit tests to run against the wrong version of CVAT
There seems to be a bug somewhere in the Docker ecosystem (it's probably
either Docker Compose, Docker Buildx or BuildKit) that causes `docker
compose build` to ignore base images that are already present in the system,
and instead fetch them from Docker Hub, if there's a custom Buildx builder
configured. There's a bug report here:
<https://github.com/docker/compose/issues/9939>.

This bug means that when the build pipeline builds the `cvat_ci` image, it's
based on the latest release of `cvat/server` from Docker Hub instead of the
version that we just built. Consequently, we run the unit tests against that
release instead of the development version.

Fortunately, we don't actually need to set up a Buildx builder in most jobs
(including the `unit_testing` job), so just don't do that.

Also, use `cvat/server:local` as the base image in `Dockerfile.ci`.
This will prevent a similar bug from reoccurring in the future, since the
`local` tag should never be uploaded to Docker Hub.