SDK: Remove redundant README templates
These templates are used for the generation of `cvat-sdk/README.md`, which
is moved into `cvat-sdk/docs/README.md`, which is then not used for
anything. It's not included in the online docs (its role there is served by
`site/content/en/docs/api_sdk/sdk/_index.md`), and it's not included in the
sdist (`cvat-sdk/gen/templates/README.md.template` is the README included
in those). So it's just a redundant documentation page.