Fix the documentation build
In 87f07fd3, I thought that the extra README that eventually ended up in
`cvat-sdk/docs/README.md` was unused, which turned out to be not quite true.
One section from it _was_ used to generate the online API index page.
Since I removed the file, the online docs now fail to build.

Rather than restore the entire README, restore just that one section as a
standalone file. This achieves the same result as before, but with simpler
code.