server-proxy.ts: make `login` do JSON requests instead of form requests
As far as I can see, this is the only request that the UI makes in the
`application/x-www-form-urlencoded` format, and there's no reason for it to
do that. Switching to JSON makes the code more readable and consistent with
other code, and enables us to retire `application/x-www-form-urlencoded`
support in the server, if we decide to do so.