Add the API schema to the repository
Instead of generating the schema during the SDK build, store it in the
repository. Add a CI check that verifies that the schema matches the server.

This has the following benefits:

* It becomes easy to see if and how a pull request affects the API
  (syntactically, at least). Since the schema is generated dynamically by
  drf-spectacular, it's not always obvious what effects a code change will
  have on the declared API. By requiring PRs to explicitly change the
  schema, we enable such changes to be reviewed directly.

* It removes one step from the SDK generation procedure, making it easier for
  people to contribute. The schema generation step was, IMO, the most
  annoying (because you have to build the server) and the most error-prone
  (because it's easy to use the wrong version of the server).

There's also a drawback: there's now one more step to change the API,
updating the schema. I think this is acceptable, because if you changed the API,
then to test your modifications you probably started the new version of the server,
so you can just get the schema from there. To make it easier, make the CI provide
the expected version of the schema, the diff from the current version, and the command
to regenerate the schema locally.