Update the developer's guide to show how to install production dependencies
This is needed to regenerate the API schema. It could also be useful just to
be able to test the production configuration.

Note that `python-lda` is not a real Arch package. I assume `python-ldap` is
what was actually meant, but there's no point in installing it, because a
new copy will be installed by pip.