Fix most discrepancies between the different ways of generating the API schema (#5889)
Currently, if you try to generate the API schema with the VS Code launch
configuration, it'll remove the `minimum` and `maximum` attributes from
the definition of some of the integer fields. This is because those
attributes come from the configured database engine in use. The
"canonical" version of the schema is generated from the Docker image,
which uses the production configuration, and thus PostgreSQL; while
using the VS Code launch configuration will default to the development
configuration, and thus SQLite.

Fix it by aligning the configurations. Since ultimately, the schema is
for the production version of the server, override the configuration to
production for the VS Code launch configuration.

Note that there's one more discrepancy that remains; that of the API
version number. I haven't figured out the best way to fix that yet.