Git sync fix (#5776)
Currently when user try to open the task with git repository connected
he will see this (screenshot)

![2023-02-27_17-01](https://user-images.githubusercontent.com/57707673/221584442-329c8592-6524-4715-9c76-7631553dc135.png)
This happens after first synchronization and reloading the page. The
cause of this is ssh-agent, that is not running in cvat_server
container.
This PR will enable ssh-agent on cvat_server so this issue disappears.