Dockerfile: remove unnecessary packages
`build-essential` contains tools for building Debian packages. We don't need
those.

The `libgeos` stuff is a bit weird. It was included on purpose in 751937bf,
because the image apparently didn't run without it, but I was unable to
reproduce this problem. I can see that Shapely uses this library, but the
Shapely wheel bundles its own copy. It's also used by
`django.contrib.gis.geos`, but as far as I can tell, neither we nor any of
our dependencies import this package.

I can only conclude that whatever reasons existed to include this library
previously no longer apply.