Dockerfile: build PyAV and its dependencies in a separate stage
This way, it can be done in parallel with pip fetching all other packages in
the main build stage. In practice, I find that the build of PyAV finishes
before pip is done downloading, so we basically get it done for free (in
terms of time).

With this change, I measured a build time of 9:09 (starting from scratch).