Stop adding package source code into the release Docker images (#6040)
The policy that mandated this is no longer relevant now that CVAT is no
longer developed by Intel. Moreover, the source code included was not
even complete (it didn't contain Python or NPM packages).

This saves ~1.6 GB in the unpacked image (and probably a bunch of build
time too, but I didn't measure it).