Rename requirements files to `*.in` to prepare for `pip-compile-multi` integration
For `dataset_manifest`, direct the users to use the `.in` file instead of
the `.txt` file (which I will soon recreate). I don't want users to use a
file with pinned versions, because those are specific to the Python
configuration (version, OS), and people might use a variety of
configurations. Using the `.in` file preserves the same experience that users
have now.

I don't really want to create a pinned version file for `dataset_manifest` at all,
but `pip-compile-multi` forces me to.