Loosen Python dependency requirements and upgrade dependencies
Now that we're using pip-compile-multi, we don't have to strictly pin our
dependency versions in the `.in` files. We can define loose version
requirements and let PCM find appropriate upgrades when we want.

Make use of that and loosen up most of the version constraints in `base.in`
and `production.in` (`testing.in` and `development.in` aren't as important,
so I'll deal with them some other time). Then upgrade dependencies within
the new constraints.

This fixes some of the issues reported by Snyk (vulnerabilities in sqlparse
and NumPy).

Some packages listed in `base.in` were used only transitively, or seemingly
just not at all. Remove those entirely.

Don't upgrade some of the packages, because I feel that it could cause more
breakages that I currently have time to deal with. I'll come back to them
another time.