Make serverless function naming and Docker image naming consistent
For function names, take the relative path, lowercase, join with dashes and
replace underscores with dashes.

For image names, take the relative path, lowercase and join with dots.

In a couple cases, rename the function directory instead of fixing the
config.

SAM still violates the convention; keep it as-is for now, as to not break
the references from published articles.