Fix logging in the TransT serverless function (#6290)
The Docker image for this function includes Conda, and uses `conda run`
to run the Nuclio processor. Unfortunately, `conda run` buffers the
entire output of the child process until that process exits, and since
the processor never exits, its logs are never printed (and slowly
consume memory).

Conda is actually completely useless here, so just get rid of it, which
fixes the problem.

In addition, apply a few other improvements:

* Synchronize the PyTorch and TransT versions between the CPU and GPU
   variants.

* Replace `opencv-python` with `opencv-python-headless`, which has fewer
   dependencies.

* Use the `ADD` command instead of wget.

Altogether, these improvements shave ~780 MB off the size of the CPU
image (didn't check for the GPU one).