CLI: don't set `discard_unknown_keys`
For unclear reasons, when this setting is enabled, the OpenAPI Generator
runtime becomes unable to parse `AnnotationsRead` values from JSON.

Specifically, it thinks that the JSON value is ambiguous, because it can be
parsed both as `LabeledData` and as `IOBase`. The former makes sense, but
the latter does not. I think it's just a bug in the runtime.

This line is why the issue only occurs when `discard_unknown_keys` is set:

https://github.com/OpenAPITools/openapi-generator/blob/v6.0.1/modules/openapi-generator/src/main/resources/python/model_utils.mustache#L1139

The CLI currently doesn't parse `AnnotationRead` anywhere, so this patch
doesn't have any user-visible effects. However, I'm prototyping a new CLI
command which will need to parse annotations.

Change `tasks_create` so that it routes kwargs to the appropriate destinations
instead of passing all of them both to `TaskWriteRequest` and `create_from_data`.
I don't think this is _really_ necessary (the extra arguments should just be ignored),
but it makes the code more illustrative.